+++ src/main/java/de/stamme/basicquests/listeners/PlayerJoinListener.java
@@ -4,7 +4,7 @@ import de.stamme.basicquests.BasicQuestsPlugin;
 import de.stamme.basicquests.Config;
 import de.stamme.basicquests.model.PlayerData;
 import de.stamme.basicquests.model.QuestPlayer;
-import de.stamme.basicquests.util.QuestsScoreBoardManager;
+import de.stamme.basicquests.util.fastboard.QuestsScoreBoardManager;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -23,7 +23,7 @@ public class PlayerJoinListener implements Listener {
 			BasicQuestsPlugin.getPlugin().getQuestPlayers().put(player.getUniqueId(), joinedPlayer);
 
 			if (Config.showScoreboardPerDefault()) {
-				QuestsScoreBoardManager.show(joinedPlayer);
+				QuestsScoreBoardManager.show(joinedPlayer, true);
 			}
 		}
 	}