+++ src/main/java/de/stamme/basicquests/listeners/PlayerQuitListener.java
@@ -3,6 +3,7 @@ package de.stamme.basicquests.listeners;
 import de.stamme.basicquests.BasicQuestsPlugin;
 import de.stamme.basicquests.model.PlayerData;
 import de.stamme.basicquests.model.QuestPlayer;
+import de.stamme.basicquests.util.fastboard.QuestsScoreBoardManager;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerQuitEvent;
@@ -22,6 +23,8 @@ public class PlayerQuitListener implements Listener {
 			BasicQuestsPlugin.log("PlayerData Saved: " + questPlayer.getName());
 		else
 			BasicQuestsPlugin.log(Level.SEVERE, "Failed to save PlayerData: " + questPlayer.getName());
+
+		QuestsScoreBoardManager.hide(questPlayer);
 		BasicQuestsPlugin.getPlugin().getQuestPlayers().remove(questPlayer.getPlayer().getUniqueId());
 	}
 }