+++ src/main/java/de/stamme/basicquests/util/metrics/MetricsService.java
@@ -1,10 +1,11 @@
-package de.stamme.basicquests.util;
+package de.stamme.basicquests.util.metrics;
 
 import de.stamme.basicquests.BasicQuestsPlugin;
 import de.stamme.basicquests.Config;
 import de.stamme.basicquests.ServerInfo;
 import de.stamme.basicquests.model.quests.Quest;
 import de.stamme.basicquests.model.quests.QuestData;
+import de.stamme.basicquests.util.StringFormatter;
 import org.bstats.bukkit.Metrics;
 
 import java.text.DecimalFormat;